--- src/assets/client_secret_685556051749-94tc9vnciflm619semreu9msgv160cvv.apps.googleusercontent.com.json
@@ -1,11 +1,0 @@
-{
-  "web": {
-    
-    "project_id": "barberapp-387501",
-    "auth_uri": "https://accounts.google.com/o/oauth2/auth",
-    "token_uri": "https://oauth2.googleapis.com/token",
-    "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
-    "client_secret": "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",
-    "javascript_origins": ["http://localhost:3000"]
-  }
-}
