+++ src/components/calendar/AgendaScreen.tsx
@@ -70,7 +70,6 @@ export const AgendaScreen = ({dateCurr, closeModal}: Props) => {
     updateTotalCost({...servicesFinal, start: startD, end: endD});
     setIsVisibleModal(true);
   };
-  
 
   return (
     <>
