+++ src/const/configurations.ts
@@ -1,0 +1,1 @@
+export const LIMITHOURSSERVICE = 2;
